refactor: moved PSF_Affine_Model to its own module
